pax-app/Wiki#186 Improve UI of credit card flux
