+++ lib/screens/home_screen/home_screen.dart
@@ -2,8 +2,7 @@ import 'package:Pax/components/base_screen/base_screen.dart';
 import 'package:Pax/components/drawer/drawer.dart';
 import 'package:Pax/screens/became_provider_screen/became_provider_tabs/info_tab.dart';
 import 'package:Pax/screens/config_screen/config_screen.dart';
-import 'package:Pax/screens/my_card_screen/add_card.dart';
-import 'package:Pax/screens/my_card_screen/my_card.dart';
+import 'package:Pax/screens/my_card_screen/selection_card.dart';
 import 'package:Pax/screens/perfil_screen/perfil_screen.dart';
 import 'package:Pax/screens/my_conversations/my_conversations.dart';
 import 'package:Pax/screens/provider_panel/provider_panel.dart';
@@ -12,7 +11,6 @@ import 'package:flutter/material.dart';
 import 'package:Pax/screens/general_categories_screen/services_general_categories.dart';
 
 import '../../components/drawer/drawer.dart';
-import '../my_card_screen/my_card.dart';
 
 class HomeScreen extends StatelessWidget {
   final _pageController = PageController();
