+++ lib/screens/my_card_screen/add_card.dart
@@ -1,12 +1,9 @@
-import 'package:Pax/components/base_screen/base_screen.dart';
+import 'package:Pax/components/app_bar/white_appbar.dart';
 import 'package:Pax/components/button%20/button.dart';
 import 'package:Pax/components/text_input/text_input.dart';
-import 'package:Pax/screens/home_screen/home_screen.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_credit_card/flutter_credit_card.dart';
-
-import 'my_card.dart';
 
 class AddCard extends StatefulWidget {
   @override
@@ -73,80 +70,40 @@ class _AddCardState extends State<AddCard> {
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-        child: Column(
-      children: <Widget>[
-        CreditCardWidget(
-          cardNumber: this._addCard.text,
-          expiryDate: this._addExpiration.text,
-          cardHolderName: this._addName.text.isEmpty
-              ? 'NOME DO TITULAR'
-              : this._addName.text.toUpperCase(),
-          cvvCode: this._addCVV.text,
-          showBackView: this._focus.hasFocus,
-          cardbgColor: Color(0xff78AA43),
-          height: 175.0,
-          textStyle: TextStyle(
-            fontFamily: 'CardFont',
-            fontSize: 16,
-            color: Color(0xff4f4f4f),
-          ),
-          width: MediaQuery.of(context).size.width,
-          animationDuration: Duration(milliseconds: 1000),
-        ),
-        Container(
+    return Scaffold(
+      resizeToAvoidBottomPadding: true,
+      appBar: WhiteAppBar('Adicinar cartão', context),
+      body: Container(
+        padding: const EdgeInsets.only(left: 20, right: 20),
+        child: SingleChildScrollView(
           child: Column(
             children: <Widget>[
-              TextInput(
-                'Número do Cartão',
-                'Número do Cartão',
-                true,
-                (String value) {
-                  return value.contains('@') ? 'Do not use the @ char.' : null;
-                },
-                TextInputType.number,
-                1,
-                focus: true,
-                controller: _addCard,
+              SizedBox(height: 15),
+              CreditCardWidget(
+                cardNumber: this._addCard.text,
+                expiryDate: this._addExpiration.text,
+                cardHolderName: this._addName.text.isEmpty
+                    ? 'NOME DO TITULAR'
+                    : this._addName.text.toUpperCase(),
+                cvvCode: this._addCVV.text,
+                showBackView: this._focus.hasFocus,
+                cardbgColor: Color(0xff78AA43),
+                height: 175.0,
+                textStyle: TextStyle(
+                  fontFamily: 'CardFont',
+                  fontSize: 16,
+                  color: Color(0xff4f4f4f),
+                ),
+                width: MediaQuery.of(context).size.width,
+                animationDuration: Duration(milliseconds: 1000),
               ),
-              TextInput(
-                'Nome do Titular',
-                'Nome do Titular',
-                true,
-                (String value) {
-                  return value.contains('@') ? 'Do not use the @ char.' : null;
-                },
-                TextInputType.text,
-                1,
-                focus: true,
-                controller: _addName,
-              ),
-              Row(
-                children: <Widget>[
-                  SizedBox(
-                    child: TextInput(
-                      'Expiração',
-                      'Expiração do Cartão',
-                      true,
-                      (String value) {
-                        return value.contains('@')
-                            ? 'Do not use the @ char.'
-                            : null;
-                      },
-                      TextInputType.number,
-                      1,
-                      focus: false,
-                      controller: _addExpiration,
-                    ),
-                    width: MediaQuery.of(context).size.width * 0.498,
-                  ),
-                  // SizedBox(
-                  //   width: MediaQuery.of(context).size.width * 0.1,
-                  // ),
-                  SizedBox(
-                    child: TextInput(
-                      'CVV',
-                      'CVV do Cartão',
+              SizedBox(height: 8),
+              Container(
+                child: Column(
+                  children: <Widget>[
+                    TextInput(
+                      'Número do Cartão',
+                      'Número do Cartão',
                       true,
                       (String value) {
                         return value.contains('@')
@@ -156,36 +113,83 @@ class _AddCardState extends State<AddCard> {
                       TextInputType.number,
                       1,
                       focus: true,
-                      controller: _addCVV,
-                      focusNd: _focus,
+                      controller: _addCard,
                     ),
-                    width: MediaQuery.of(context).size.width * 0.39,
-                  )
-                ],
-              )
+                    TextInput(
+                      'Nome do Titular',
+                      'Nome do Titular',
+                      true,
+                      (String value) {
+                        return value.contains('@')
+                            ? 'Do not use the @ char.'
+                            : null;
+                      },
+                      TextInputType.text,
+                      1,
+                      focus: true,
+                      controller: _addName,
+                    ),
+                    Row(
+                      children: <Widget>[
+                        SizedBox(
+                          child: TextInput(
+                            'Expiração',
+                            'Expiração do Cartão',
+                            true,
+                            (String value) {
+                              return value.contains('@')
+                                  ? 'Do not use the @ char.'
+                                  : null;
+                            },
+                            TextInputType.number,
+                            1,
+                            focus: false,
+                            controller: _addExpiration,
+                          ),
+                          width: MediaQuery.of(context).size.width * 0.498,
+                        ),
+                        SizedBox(
+                          child: TextInput(
+                            'CVV',
+                            'CVV do Cartão',
+                            true,
+                            (String value) {
+                              return value.contains('@')
+                                  ? 'Do not use the @ char.'
+                                  : null;
+                            },
+                            TextInputType.number,
+                            1,
+                            focus: true,
+                            controller: _addCVV,
+                            focusNd: _focus,
+                          ),
+                          width: MediaQuery.of(context).size.width * 0.39,
+                        )
+                      ],
+                    )
+                  ],
+                ),
+              ),
+              SizedBox(height: 16),
+              Button(
+                buttonText: 'Adicionar',
+                type: 'default',
+                tapHandler: _addCard.text.length == 0 ||
+                        _addCVV.text.length == 0 ||
+                        _addExpiration.text.length == 0 ||
+                        _addName.text.length == 0
+                    ? null
+                    : () {
+                        _showDialog();
+                      },
+                isSmall: false,
+              ),
+              SizedBox(height: 15.0),
             ],
           ),
         ),
-        SizedBox(
-          height: 5.0,
-        ),
-        Button(
-          buttonText: 'Adicionar Cartão',
-          type: 'default',
-          tapHandler: _addCard.text.length == 0 ||
-                  _addCVV.text.length == 0 ||
-                  _addExpiration.text.length == 0 ||
-                  _addName.text.length == 0
-              ? null
-              : () {
-                  _showDialog();
-                },
-          isSmall: false,
-        ),
-        SizedBox(
-          height: 5.0,
-        ),
-      ],
-    ));
+      ),
+    );
   }
 }
