+++ lib/theme/colors.dart
@@ -7,7 +7,7 @@ final secondaryColor = const Color(0xff78AA43);
 final secondaryColorLight = const Color(0xff7CB342);
 final secondaryColorDimmed = const Color(0x6678AA43);
 final primaryButtonShadow = const Color(0x8778AA43);
-final backgroundColor = const Color(0xfff3f3f3);
+final backgroundColor = const Color(0xfff1f1f1);
 final errorColor = const Color(0xffFF4B4B);
 final errorColorLight = const Color(0x66FF4B4B);
 final colorWhite = const Color(0xffffffff);
